refactor(app): extract useTabNavigation + isFocusInTerminalPanel
Phase 1 step 6 (partial) of the god-file cleanup plan. Move the pure
tab-and-sub-tab navigation primitives out of App.tsx:

- src/hooks/useTabNavigation.ts: nextTab, jumpToTab, moveActiveTab,
  nextShellSubTab, jumpToShellSubTab, moveActiveShellSubTab.
- src/utils/focus.ts: isFocusInTerminalPanel (DOM query, used by
  Cmd+T routing and the keyboard shortcut handler).

The heavier tab-lifecycle functions (setActiveTab, newTab, newShellTab,
closeTab, closeTabNow, pushClosedTab, reopenLastClosedTab,
autoRestoreDiscoveredSessions, dispatchTerminalReplay,
applyShareModeToTab, updateTab, updateActiveTab) stay in App.tsx for
now — they have wider dependencies (project bucket swap, model picker
recompute, terminal replay, shell consent prompts) that would benefit
from landing alongside one another in a follow-up extraction.

App.tsx 5678 -> 5568 lines. tsc clean, eslint clean, 305/305 tests
pass. No behavior change.
